pushing final commit to new repo

+++ src/main/java/com/lmig/moviedb/MovieRestController.java
@@ -35,95 +35,38 @@ public class MovieRestController {
 	
 	@Autowired
 	private UserRepository userRepository;
+
 	
-//TODO do we need this route anymore? 	
-	@RequestMapping(path = "/api/Movie", method = RequestMethod.GET)
-	public Movie movie(String movie, Integer year, String genre, String rating, Integer score, String language,
-			String person) {
-		if (year == null) {
-			year = 2000;
-		}
-		if (score == null) {
-			score = 2;
-		}
-		return new Movie(movie, year, genre, rating, score, language, person);
-	}
-	
-	//TODO do we need this route anymore? 
-//	@RequestMapping(path = "/api/Person", method = RequestMethod.GET)
-//	public ArrayList<Person> person(String name, String birthPlace, String gender, Integer popScore) {
-//		if (popScore == null) {
-//			popScore = 1;
-//		}
-//		ArrayList<Person> personArray = new ArrayList<Person>();
-//		ArrayList<Person> resultArray = new ArrayList<Person>();
-//		;
-//		System.out.println("ok1");
-		// String about="";
-//		personArray.add(new Person("ravi", "indy", "male", 5, "ravi is going to gfc"));
-//		personArray.add(new Person("ravi", "Seattle", "male", 5, "ravi is going to gfc"));
-//		personArray.add(new Person("David", "Dover", "male", 5, "David is going to gfc"));
-//		personArray.add(new Person("Tracilyn", "indy", "Female", 5, "Tracilyn is going to gfc"));
-//		System.out.println("ok2");
-//		for (int i = 0; i < personArray.size(); i++) {
-//			if (personArray.get(i).name.equals(name)) {
-//				System.out.println("ok3");
-//				resultArray.add(personArray.get(i));
-//			}
-//
-//		}
-//		System.out.println(resultArray.toString());
-//		return resultArray;
-//	}
-
-//	TODO do we need this route anymore?
-	@RequestMapping(value = "/api/AddPerson", method = RequestMethod.POST)
-	public ArrayList<Person> add(@RequestBody Person person, String personName) {
-
-		ArrayList<Person> personArray = new ArrayList<Person>();
-		ArrayList<Person> resultArray = new ArrayList<Person>();
-
-		System.out.println("ok1");
-		// String about="";
-//		personArray.add(new Person("ravi", "Seattle", "male", 5, "ravi is going to gfc"));
-//		personArray.add(new Person("David", "Dover", "male", 5, "David is going to gfc"));
-//		personArray.add(new Person("Tracilyn", "indy", "Female", 5, "Tracilyn is going to gfc"));
-//		personArray.add(new Person(person.getName(), "indy", "male", 5, "ravi is going to gfc"));
-//		personArray.add(new Person(personName, "indy", "male", 5, "ravi is going to gfc"));
-		System.out.println(resultArray.toString());
-		return personArray;
-	}
-
 	@RequestMapping(value = "/api/addPerson", method = RequestMethod.POST)
-	public void addPerson(@RequestBody Person person) {
+	public HttpStatus addPerson(@RequestBody Person person) {
+		if (person == null) {
+			 return HttpStatus.BAD_REQUEST;
+		}
 		personRepository.save(person);
+		return HttpStatus.OK;
 	}
 	
 	@RequestMapping(value = "/api/addUser", method = RequestMethod.POST)
-	public void addUser(@RequestBody User user) {
+	public HttpStatus addUser(@RequestBody User user) {
+		if (user == null) {
+			 return HttpStatus.BAD_REQUEST;
+		}
 		userRepository.save(user);
+		return HttpStatus.OK;
 	}
 
 	@RequestMapping(value = "/api/addMovie", method = RequestMethod.POST)
 //	added the @ResponseBody as part of JUnit testing
 	@ResponseBody
-	public void addMovie(@RequestBody Movie movie) {
+	public HttpStatus addMovie(@RequestBody Movie movie) {
+		if (movie == null) {
+			 return HttpStatus.BAD_REQUEST;
+		}
+		System.out.println("ok from JUnit");
 		movieRepository.save(movie);
-
-		// personArray.add(new Person(person.getName(), "indy", "male", 5, "ravi
-		// is going to gfc"));
-		// personArray.add(new Person(personName, "indy", "male", 5, "ravi is
-		// going to gfc"));
-		// System.out.println(resultArray.toString());
-		// return personArray;
+		return HttpStatus.OK;
 	}
-//old version of putmovie
-//	@RequestMapping(value = "/api/putMovie", method = RequestMethod.PUT)
-//	public void updateMovie(@RequestBody Movie movie) {
-//		movieRepository.save(movie);
-//		// post man URL for putting it is http://localhost:8080/api/putMovie/.
-//		// Input JSON if id is existing, it will get updated.
-//	}
+
 	
     @RequestMapping(path = "/api/putMovie", method = RequestMethod.PUT)
     public ResponseEntity<Movie> updateMovie(@RequestBody Movie m) {  
@@ -163,33 +106,27 @@ public class MovieRestController {
         }
         System.out.println(userRepository.findById(u.getId()));
         User existing = userRepository.findById(u.getId());
+        if (existing == null) {
+			return new ResponseEntity<User>(existing, HttpStatus.NOT_FOUND);
+		}
         System.out.println(existing.name);
 //        existing.name="david";
         existing.merge(u);
         userRepository.save(existing);
+        System.out.println("");
         System.out.println("ok 2");
         return new ResponseEntity<User>(existing, HttpStatus.OK);
     }
     
-//	@RequestMapping(value = "/api/putPerson", method = RequestMethod.PUT)
-//	public void updatePerson(@RequestBody Person person) {
-//		personRepository.save(person);
-//		// post man URL for putting it is http://localhost:8080/api/putPerson/.
-//		// Input JSON if id is existing, it will get updated.
-//	}
-
-//   old code
-//	@RequestMapping(value = "/api/putUser", method = RequestMethod.PUT)
-//	public void updateUser(@RequestBody User user) {
-//		userRepository.save(user);
-//		// post man URL for putting it is http://localhost:8080/api/putUser/.
-//		// Input JSON if id is existing, it will get updated.
-//	}
-	
+    
 	// this deletes single movie using Pathvariable
 	@RequestMapping(value = "/api/deleteMovie/{id}", method = RequestMethod.DELETE)
-	public void deleteMovie(@PathVariable(name = "id", required = true) int id) {
+	public HttpStatus deleteMovie(@PathVariable(name = "id", required = true) int id) {
+		if (id == 0) {
+			return HttpStatus.BAD_REQUEST;
+		}
 		movieRepository.delete(id);
+			return HttpStatus.OK;
 		// post man URL for putting it is
 		// http://localhost:8080/api/deleteMovie/98.
 		// Anything in the body will get ignored
@@ -197,14 +134,13 @@ public class MovieRestController {
 	}
 
 	// this deletes single movie using delete method and data supplied thru body
-	@RequestMapping(value = "/api/deleteMovie", method = RequestMethod.DELETE)
-	public void deleteMovie(@RequestBody Integer id) {
-		movieRepository.delete(id);
+//	@RequestMapping(value = "/api/deleteMovie", method = RequestMethod.DELETE)
+//	public void deleteMovie(@RequestBody Integer id) {
+//		movieRepository.delete(id);
 		// post man URL for putting it is
 		// http://localhost:8080/api/deleteMovie/98.
 		// Anything in the body will get ignored
-
-	}
+//	}
 
 	// multiple movies deleted supplied in the form of a list from the postman
 	// body
@@ -212,14 +148,14 @@ public class MovieRestController {
 	// body is supplied
 	// in the list form as [1,2] where 1 and 2 are the IDs that we want to pass
 	// to the list
-	@RequestMapping(value = "/api/deleteMovies", method = RequestMethod.DELETE)
-	public void deleteMovie(@RequestBody List<Integer> ids) {
-		for (Iterator iterator = ids.iterator(); iterator.hasNext();) {
-			Integer integer = (Integer) iterator.next();
-			movieRepository.delete(integer);
-		}
-
-	}
+//	@RequestMapping(value = "/api/deleteMovies", method = RequestMethod.DELETE)
+//	public void deleteMovie(@RequestBody List<Integer> ids) {
+//		for (Iterator iterator = ids.iterator(); iterator.hasNext();) {
+//			Integer integer = (Integer) iterator.next();
+//			movieRepository.delete(integer);
+//		}
+//
+//	}
 
 	// this deletes single Person using Pathvariable
 	@RequestMapping(value = "/api/deletePerson/{id}", method = RequestMethod.DELETE)
